requirements.txt
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-bigquery==3.30.0
google-cloud-core==2.4.3
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.70.0
grpcio-status==1.70.0
idna==3.10
numpy==1.24.4
packaging==24.2
pandas==2.0.3
pandas-gbq>=0.26.1
proto-plus==1.26.1
protobuf==5.29.4
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.3
rsa==4.9
six==1.17.0
tzdata==2025.2
urllib3==2.2.3
